AF-94 Fixed issue with leftover app timeouts and btw added VMX raw disk (#108)
Two changes here:
* After a night of testing I found the app timeout is not working correctly - it's just creating the state not checking if the app exists, so bloating the database. Added a quick fix for that and now removing app timeout on deallocation. Updated compactdb test to check this behavior as well. Also found deadlock with channels there - oh those channels...
* Also today we found that git >=2.35 don't really like to be running on windows with wrong owner for the workspace, so I've added VMX driver ability to create raw disks instead of formatting them on the host system to the limited amount of filesystems. That was a relatively legacy feature, so with modern bait images containing diskmount role (which is formatting the attached raw disks) we don't need to strictly rely on host OS for that and can easily format the disk during the VM init procedures.

**WARNING:** VMX driver disks default was changed - before type of disk by default was "extfs", now it's "raw" and you need to change your labels to use `type: "extfs"` if you did not defined that or defined as empty string value.

+++ tests/helper/fish.go
@@ -65,7 +65,12 @@ func NewAfInstance(tb testing.TB, name, cfg string) *AFInstance {
 		waitForLog: make(map[string]func(string, string) bool),
 	}
 
-	afi.workspace = tb.TempDir()
+	// Not using here tb.TempDir to have an ability to save on cleanup for investigation
+	var err error
+	if afi.workspace, err = os.MkdirTemp("", "fish"); err != nil {
+		tb.Fatal("INFO: Unable to create workspace:", afi.nodeName, err)
+		return nil
+	}
 	tb.Log("INFO: Created workspace:", afi.nodeName, afi.workspace)
 
 	cfg += fmt.Sprintf("\nnode_name: %q", afi.nodeName)
@@ -146,6 +151,11 @@ func (afi *AFInstance) Cleanup(tb testing.TB) {
 	tb.Helper()
 	tb.Log("INFO: Cleaning up:", afi.nodeName, afi.workspace)
 	afi.Stop(tb)
+
+	if tb.Failed() {
+		tb.Log("INFO: Keeping workspace for checking:", afi.workspace)
+		return
+	}
 	os.RemoveAll(afi.workspace)
 }
